fix: Resolve clippy warnings
- Add #[allow(clippy::too_many_arguments)] to functions with 8 parameters
- Change Into<T> implementation to From<T> (clippy::from_over_into)
- Remove unnecessary double cast (u32 as u32 as f64 -> u32 as f64)

All clippy checks now pass with -D warnings.

+++ bindings/rust/src/provider.rs
@@ -249,7 +249,7 @@ impl Provider {
     pub fn set_pixel_format(&mut self, format: PixelFormat) -> Result<()> {
         self.set_property(
             PropertyName::PixelFormatOutput,
-            format.to_c_enum() as u32 as f64,
+            format.to_c_enum() as f64,
         )
     }
